fix: Stack.Protected 활용에 따른 자동로그인 로직 수정
